IOMP_Final/learning_engine/requirements.txt
streamlit
sqlite3
groq
python-dotenv
requests
